nuget code cleanup
