+++ nuget/src/StackExchange.StacksIcons.csproj
@@ -20,9 +20,14 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<!-- Expose `internal` items to the test assembly -->
+		<AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+			<_Parameter1>StackExchange.StacksIcons.Test</_Parameter1>
+		</AssemblyAttribute>
 	</ItemGroup>
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.Html.Abstractions" Version="2.2.0" />
+		<PackageReference Include="System.Memory" Version="4.5.5" />
 	</ItemGroup>
 </Project>
